+++ src/communities/Community.module.css
@@ -152,11 +152,13 @@ body {
   background-color: #f0f2f5;
   border-radius: 8px;
   padding: 15px;
+  width: 300px;
   margin-bottom: 20px;
   box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
 }
 
 .ranking h2 {
+  text-align: center;
   margin-top: 30px;
   color: #1877f2;
 }
