+++ src/communities/CommunityDetail.module.css
